services/driver-service/pyproject.toml
[project]
name = "driver-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Oviemena",email = "okpowoduoviemena999@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9, <4"
dependencies = [
    "flask>=3.1.0",
    "redis>=3.0.0,<4.0.0",  # Changed to be compatible with redis-py-cluster
    "redis-py-cluster>=2.1.3",
    "firebase-admin>=6.6.0",
    "requests>=2.32.3",
    "numpy>=1.24.0",  # Using older version for Python 3.9 compatibility
    "prometheus-client>=0.21.1",
    "tenacity>=9.0.0",
    "flask-jwt-extended>=4.7.1",
    "python-decouple (>=3.8,<4.0)",
]

[tool.poetry]
packages = [
    { include = "driver_service", from = "src" },
    { include = "connections", from = "../shared-utils" }
]
[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"